+++ src/main/java/de/samply/db/repository/ProjectDocumentRepository.java
@@ -2,6 +2,7 @@ package de.samply.db.repository;
 
 import de.samply.db.model.Project;
 import de.samply.db.model.ProjectDocument;
+import de.samply.document.DocumentType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -12,5 +13,6 @@ public interface ProjectDocumentRepository extends JpaRepository<ProjectDocument
 
     Optional<ProjectDocument> findFirstByProjectAndBridgeheadAndOriginalFilename(Project project, String bridgehead, String filename);
     Optional<ProjectDocument> findFirstByProjectAndOriginalFilename(Project project, String filename);
+    Optional<ProjectDocument> findFirstByProjectAndDocumentTypeAndBridgeheadOrderByCreatedAtDesc(Project project, DocumentType documentType, String bridgehead);
 
 }